Use git over https, rather than git over ssh
It seems like in recent times (see [1]), github has slowly phased out the unauthenticated `git://`/ssh protocol, and thus `git@...` links only work if you are already logged into GitHub.

This commit updates the submodules (and the README) to use the https protocol so that cloning the repositories without logging into GitHub works.

Thanks to the artifact reviewers for pointing this out!

[1] https://github.blog/2021-09-01-improving-git-protocol-security-github/